+++ src/main/java/cn/yanweijia/slimming/service/IUserService.java
@@ -3,6 +3,6 @@ package cn.yanweijia.slimming.service;
 import cn.yanweijia.slimming.model.User;
 
 public interface IUserService {
-    public User selectUser(long userId);
+    public User selectUser(Integer userId);
 
 }
